Add files via upload
My final project from my first Computer science class CS1. This is a menu driven program using a single dimensional array containing structs using a for loop to prompt the user to enter data and after is brought back to the main menu. I plan to include more features later on such as saving the vector to a text file so that data can be moved out of ram that way the program can be more useful if I decide to modify it and use it for other things.